Link para postagem de newsletter implantado
close #75
