+++ mcfields/django_assertions.py
@@ -2,5 +2,6 @@ from django.test import TestCase
 
 _test_case = TestCase()
 assert_contains = _test_case.assertContains
+assert_not_contains = _test_case.assertNotContains
 assert_true = _test_case.assertTrue
 assert_false = _test_case.assertFalse
