+++ mcfields/newsletter/tests/test_indice_newsletter.py
@@ -1,8 +1,10 @@
 import pytest
+from django.contrib.auth.models import Permission
+from django.contrib.contenttypes.models import ContentType
 from django.urls import reverse
 from model_bakery import baker
 from datetime import date
-from mcfields.django_assertions import assert_contains
+from mcfields.django_assertions import assert_contains, assert_not_contains
 from mcfields.newsletter.models import Newsletter
 
 
@@ -26,6 +28,15 @@ def resp_indice_newsletters(client, newsletters):
     return response
 
 
+# @pytest.fixture
+# def resp_indice_newsletters_usuario_logado(client_usuario_logado):
+#     """
+#     Cria uma requisição na página de índice de newsletters com usuário logado.
+#     """
+#     response = client_usuario_logado.get(reverse('newsletter:indice_newsletters'))
+#     return response
+
+
 def test_status_code_indice_newsletter(resp_indice_newsletters):
     """
     Certifica de que a página de índice de newsletters foi carregada com sucesso.
@@ -45,3 +56,23 @@ def test_dados_newsletter_indice(resp_indice_newsletters, newsletters):
         assert_contains(resp_indice_newsletters, f'<div class="newsletter-date">{pub_date}</div>')
         assert_contains(resp_indice_newsletters, f'<a href="{newsletter.get_absolute_url()}" '
                                                  f'class="newsletter-box-link">')
+
+
+def test_botao_nova_news_disponivel(usuario_senha_plana, client_usuario_logado):
+    """
+    Certifica de que, com o usuário logado e tendo permissão de postar, o botão de nova newsletter é exibido.
+    """
+    content_type = ContentType.objects.get_for_model(Newsletter)
+    permission = Permission.objects.get(codename='add_newsletter', content_type=content_type)
+    usuario_senha_plana.user_permissions.add(permission)
+    response = client_usuario_logado.get(reverse('newsletter:indice_newsletters'))
+    assert_contains(response, f'<a id="link-nova-pub" href="{reverse("newsletter:post")}">Nova publicação</a>')
+
+
+def test_botao_nova_news_indisponivel(usuario_senha_plana, client_usuario_logado):
+    """
+    Certifica de que, mesmo com o usuário logado, o botão de nova
+    newsletter não é exibido se o usuário não tem permissão.
+    """
+    response = client_usuario_logado.get(reverse('newsletter:indice_newsletters'))
+    assert_not_contains(response, 'Nova publicação')
